docs/samples/auth_api/requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
ansible==8.2.0
ansible-core==2.15.2
anyio==3.7.1
appnope==0.1.3
argcomplete==3.0.8
asgiref==3.7.1
asttokens==2.2.1
awscli==1.27.160
awscli-local==0.20
backcall==0.2.0
black==23.12.1
boto3==1.26.160
botocore==1.29.160
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
colorama==0.4.4
commitizen==3.2.1
cryptography==40.0.2
decli==0.6.0
decorator==5.1.1
distlib==0.3.6
Django==4.2.7
docutils==0.16
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.0
future @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.0.7
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
localstack-client==2.2
macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic_core==2.3.0
Pygments==2.15.1
PyJWT==2.7.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-whois==0.8.0
pytz==2023.3
PyYAML==5.4.1
qrcode==7.4.2
questionary==1.10.0
resolvelib==1.0.1
rsa==4.7.2
s3transfer==0.6.1
six @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
termcolor==2.3.0
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
vboxapi @ file:///Applications/VirtualBox.app/Contents/MacOS/sdk/installer
virtualbox==2.1.1
virtualenv==20.23.0
wcwidth==0.2.6
zipp==3.15.0
